refactor: split up livekit-data-stream crate into multiple smaller modules
