+++ livekit-data-stream/src/outgoing/mod.rs
@@ -1,4 +1,4 @@
-// Copyright 2025 LiveKit, Inc.
+// Copyright 2026 LiveKit, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,11 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use super::{
-    create_random_uuid, ByteStreamInfo, OperationType, SendError, StreamError, StreamProgress,
-    StreamResult, TextStreamInfo,
-};
-use crate::utf8_chunk::Utf8AwareChunkExt;
 use bmrng::unbounded::{UnboundedRequestReceiver, UnboundedRequestSender};
 use chrono::Utc;
 use livekit_common::{
@@ -26,276 +21,23 @@ use livekit_common::{
 use livekit_protocol as proto;
 use proto::data_stream::CompressionType;
 use std::{collections::HashMap, io::Write, path::Path, sync::Arc};
-use tokio::{io::AsyncReadExt, sync::Mutex};
+use tokio::sync::Mutex;
 
-/// Writer for an open data stream.
-pub trait StreamWriter<'a> {
-    /// Type of input this writer accepts.
-    type Input: 'a;
+use crate::utf8_chunk::Utf8AwareChunkExt;
+use crate::info::{OperationType, TextStreamInfo, ByteStreamInfo};
+use crate::utils::{StreamResult, StreamError, SendError};
 
-    /// Information about the underlying data stream.
-    type Info;
+mod stream_writer;
+pub use stream_writer::{StreamWriter, ByteStreamWriter, TextStreamWriter};
 
-    /// Returns a reference to the stream info.
-    fn info(&self) -> &Self::Info;
+mod constants;
 
-    /// Writes to the stream.
-    fn write(
-        &self,
-        input: Self::Input,
-    ) -> impl std::future::Future<Output = StreamResult<()>> + Send;
+mod raw_stream;
+use raw_stream::{RawStream, RawStreamOpenOptions};
 
-    /// Closes the stream normally.
-    fn close(self) -> impl std::future::Future<Output = StreamResult<()>> + Send;
-
-    /// Closes the stream abnormally, specifying the reason for closure.
-    fn close_with_reason(
-        self,
-        reason: &str,
-    ) -> impl std::future::Future<Output = StreamResult<()>> + Send;
-}
-
-#[derive(Clone)]
-/// Writer for an open byte data stream.
-pub struct ByteStreamWriter {
-    info: Arc<ByteStreamInfo>,
-    stream: Arc<Mutex<RawStream>>,
-}
-
-#[derive(Clone)]
-/// Writer for an open text data stream.
-pub struct TextStreamWriter {
-    info: Arc<TextStreamInfo>,
-    stream: Arc<Mutex<RawStream>>,
-}
-
-impl<'a> StreamWriter<'a> for ByteStreamWriter {
-    type Input = &'a [u8];
-    type Info = ByteStreamInfo;
-
-    fn info(&self) -> &Self::Info {
-        &self.info
-    }
-
-    async fn write(&self, bytes: &'a [u8]) -> StreamResult<()> {
-        let mut stream = self.stream.lock().await;
-        for chunk in bytes.chunks(STREAM_CHUNK_SIZE_BYTES) {
-            stream.write_chunk(chunk).await?;
-        }
-        Ok(())
-    }
-
-    async fn close(self) -> StreamResult<()> {
-        self.stream.lock().await.close(None).await
-    }
-
-    async fn close_with_reason(self, reason: &str) -> StreamResult<()> {
-        self.stream.lock().await.close(Some(reason)).await
-    }
-}
-
-impl<'a> StreamWriter<'a> for TextStreamWriter {
-    type Input = &'a str;
-    type Info = TextStreamInfo;
-
-    fn info(&self) -> &Self::Info {
-        &self.info
-    }
-
-    async fn write(&self, text: &'a str) -> StreamResult<()> {
-        let mut stream = self.stream.lock().await;
-        for chunk in text.as_bytes().utf8_aware_chunks(STREAM_CHUNK_SIZE_BYTES) {
-            stream.write_chunk(chunk).await?;
-        }
-        Ok(())
-    }
-
-    async fn close(self) -> StreamResult<()> {
-        self.stream.lock().await.close(None).await
-    }
-
-    async fn close_with_reason(self, reason: &str) -> StreamResult<()> {
-        self.stream.lock().await.close(Some(reason)).await
-    }
-}
-
-struct RawStreamOpenOptions {
-    header: proto::data_stream::Header,
-    destination_identities: Vec<ParticipantIdentity>,
-    packet_tx: UnboundedRequestSender<proto::DataPacket, Result<(), SendError>>,
-}
-
-struct RawStream {
-    id: String,
-    progress: StreamProgress,
-    is_closed: bool,
-    /// Request channel for sending packets.
-    packet_tx: UnboundedRequestSender<proto::DataPacket, Result<(), SendError>>,
-}
-
-impl RawStream {
-    async fn open(options: RawStreamOpenOptions) -> StreamResult<Self> {
-        let id = options.header.stream_id.to_string();
-        let bytes_total = options.header.total_length;
-
-        let packet = Self::create_header_packet(options.header, options.destination_identities);
-        Self::send_packet(&options.packet_tx, packet).await?;
-
-        Ok(Self {
-            id,
-            progress: StreamProgress { bytes_total, ..Default::default() },
-            is_closed: false,
-            packet_tx: options.packet_tx,
-        })
-    }
-
-    async fn write_chunk(&mut self, bytes: &[u8]) -> StreamResult<()> {
-        let packet = Self::create_chunk_packet(&self.id, self.progress.chunk_index, bytes);
-        Self::send_packet(&self.packet_tx, packet).await?;
-        self.progress.bytes_processed += bytes.len() as u64;
-        self.progress.chunk_index += 1;
-        Ok(())
-    }
-
-    /// Writes opaque bytes split into MTU-sized chunks on raw byte boundaries.
-    ///
-    /// Used for byte payloads and for compressed (deflate-raw) content, where the bytes
-    /// are opaque and must not be split on UTF-8 boundaries.
-    async fn write_raw_chunks(&mut self, bytes: &[u8]) -> StreamResult<()> {
-        for chunk in bytes.chunks(STREAM_CHUNK_SIZE_BYTES) {
-            self.write_chunk(chunk).await?;
-        }
-        Ok(())
-    }
-
-    /// Streams a file's contents into MTU-sized chunks, optionally deflate-raw compressing
-    /// on the fly. The whole file is never buffered in memory at once.
-    async fn write_file(&mut self, path: impl AsRef<Path>, compress: bool) -> StreamResult<()> {
-        let mut file = tokio::fs::File::open(path).await?;
-        let mut read_buf = vec![0u8; 8192];
-
-        if compress {
-            let mut encoder =
-                flate2::write::DeflateEncoder::new(Vec::new(), flate2::Compression::default());
-            loop {
-                let n = file.read(&mut read_buf).await?;
-                if n == 0 {
-                    break;
-                }
-                // Writing into a `Vec` is infallible.
-                encoder.write_all(&read_buf[..n]).expect("deflate write to Vec is infallible");
-                // Drain whole MTU-sized chunks of compressed output as they accumulate so
-                // we never hold the full compressed file in memory.
-                while encoder.get_ref().len() >= STREAM_CHUNK_SIZE_BYTES {
-                    let rest = encoder.get_mut().split_off(STREAM_CHUNK_SIZE_BYTES);
-                    let chunk = std::mem::replace(encoder.get_mut(), rest);
-                    self.write_chunk(&chunk).await?;
-                }
-            }
-            // Flush the final deflate block and send whatever compressed bytes remain.
-            let remaining = encoder.finish().expect("deflate finish into Vec is infallible");
-            self.write_raw_chunks(&remaining).await?;
-        } else {
-            let mut pending: Vec<u8> = Vec::new();
-            loop {
-                let n = file.read(&mut read_buf).await?;
-                if n == 0 {
-                    break;
-                }
-                pending.extend_from_slice(&read_buf[..n]);
-                while pending.len() >= STREAM_CHUNK_SIZE_BYTES {
-                    let rest = pending.split_off(STREAM_CHUNK_SIZE_BYTES);
-                    let chunk = std::mem::replace(&mut pending, rest);
-                    self.write_chunk(&chunk).await?;
-                }
-            }
-            if !pending.is_empty() {
-                self.write_chunk(&pending).await?;
-            }
-        }
-        Ok(())
-    }
-
-    async fn close(&mut self, reason: Option<&str>) -> StreamResult<()> {
-        if self.is_closed {
-            Err(StreamError::AlreadyClosed)?
-        }
-        let packet = Self::create_trailer_packet(&self.id, reason);
-        Self::send_packet(&self.packet_tx, packet).await?;
-        self.is_closed = true;
-        Ok(())
-    }
-
-    async fn send_packet(
-        tx: &UnboundedRequestSender<proto::DataPacket, Result<(), SendError>>,
-        packet: proto::DataPacket,
-    ) -> StreamResult<()> {
-        tx.send_receive(packet)
-            .await
-            .map_err(|_| StreamError::Internal)? // request channel closed
-            .map_err(|_| StreamError::SendFailed) // data channel error
-    }
-
-    fn create_header_packet(
-        header: proto::data_stream::Header,
-        destination_identities: Vec<ParticipantIdentity>,
-    ) -> proto::DataPacket {
-        proto::DataPacket {
-            kind: proto::data_packet::Kind::Reliable.into(),
-            participant_identity: String::new(), // populate later
-            destination_identities: destination_identities.into_iter().map(|id| id.0).collect(),
-            value: Some(livekit_protocol::data_packet::Value::StreamHeader(header)),
-            // TODO: placeholder for reliable data transport
-            ..Default::default()
-        }
-    }
-
-    fn create_chunk_packet(id: &str, chunk_index: u64, content: &[u8]) -> proto::DataPacket {
-        let chunk = proto::data_stream::Chunk {
-            stream_id: id.to_string(),
-            chunk_index,
-            content: content.to_vec(),
-            ..Default::default()
-        };
-        proto::DataPacket {
-            kind: proto::data_packet::Kind::Reliable.into(),
-            participant_identity: String::new(), // populate later
-            value: Some(livekit_protocol::data_packet::Value::StreamChunk(chunk)),
-            ..Default::default()
-        }
-    }
-
-    fn create_trailer_packet(id: &str, reason: Option<&str>) -> proto::DataPacket {
-        let trailer = proto::data_stream::Trailer {
-            stream_id: id.to_string(),
-            reason: reason.unwrap_or_default().to_owned(),
-            ..Default::default()
-        };
-        proto::DataPacket {
-            kind: proto::data_packet::Kind::Reliable.into(),
-            participant_identity: String::new(), // populate later
-            value: Some(livekit_protocol::data_packet::Value::StreamTrailer(trailer)),
-            ..Default::default()
-        }
-    }
-}
-
-impl Drop for RawStream {
-    /// Close stream normally if not already closed.
-    fn drop(&mut self) {
-        if self.is_closed {
-            return;
-        }
-        let packet = Self::create_trailer_packet(&self.id, None);
-        let packet_tx = self.packet_tx.clone();
-        // Use try_current() instead of assuming a Tokio runtime exists.
-        // The drop can run on a non-Tokio thread (e.g. a GC finalizer in
-        // Unity/.NET) or after the runtime has shut down, in which case
-        // we silently skip the trailer — the connection is going away anyway.
-        if let Ok(handle) = tokio::runtime::Handle::try_current() {
-            handle.spawn(async move { Self::send_packet(&packet_tx, packet).await });
-        }
-    }
+/// Generates a random stream identifier (UUID v4).
+fn create_random_uuid() -> String {
+    uuid::Uuid::new_v4().to_string()
 }
 
 /// Options used when opening an outgoing byte data stream.
@@ -356,10 +98,10 @@ impl OutgoingStreamManager {
             destination_identities: dests,
             packet_tx: self.packet_tx.clone(),
         };
-        let writer = TextStreamWriter {
-            info: Arc::new(TextStreamInfo::from_headers(header, text_header)),
-            stream: Arc::new(Mutex::new(RawStream::open(open_options).await?)),
-        };
+        let writer = TextStreamWriter::new(
+            Arc::new(TextStreamInfo::from_headers(header, text_header)),
+            Arc::new(Mutex::new(RawStream::open(open_options).await?)),
+        );
         Ok(writer)
     }
 
@@ -382,10 +124,10 @@ impl OutgoingStreamManager {
             destination_identities: dests,
             packet_tx: self.packet_tx.clone(),
         };
-        let writer = ByteStreamWriter {
-            info: Arc::new(ByteStreamInfo::from_headers(header, byte_header)),
-            stream: Arc::new(Mutex::new(RawStream::open(open_options).await?)),
-        };
+        let writer = ByteStreamWriter::new(
+            Arc::new(ByteStreamInfo::from_headers(header, byte_header)),
+            Arc::new(Mutex::new(RawStream::open(open_options).await?)),
+        );
         Ok(writer)
     }
 
@@ -448,7 +190,7 @@ impl OutgoingStreamManager {
         if should_compress {
             stream.write_raw_chunks(payload.as_compressed()?).await?;
         } else {
-            for chunk in payload.uncompressed.utf8_aware_chunks(STREAM_CHUNK_SIZE_BYTES) {
+            for chunk in payload.uncompressed.utf8_aware_chunks(constants::STREAM_CHUNK_SIZE_BYTES) {
                 stream.write_chunk(chunk).await?;
             }
         }
@@ -476,7 +218,7 @@ impl OutgoingStreamManager {
         }
         let bytes = data.as_ref();
         let stream_id = options.id.clone().unwrap_or_else(create_random_uuid);
-        let name = options.name.clone().unwrap_or_else(|| BYTE_DEFAULT_NAME.to_owned());
+        let name = options.name.clone().unwrap_or_else(|| constants::BYTE_DEFAULT_NAME.to_owned());
         let total_length = bytes.len() as u64;
         let mut payload = MaybeCompressed::new(bytes);
 
@@ -641,7 +383,7 @@ fn header_packet_fits(
 ) -> bool {
     use prost::Message;
     let packet = RawStream::create_header_packet(header.clone(), destinations.to_vec());
-    packet.encoded_len() <= STREAM_CHUNK_SIZE_BYTES
+    packet.encoded_len() <= constants::STREAM_CHUNK_SIZE_BYTES
 }
 
 /// Enforces the header-packet MTU budget on the chunked path (the inline path falls back
@@ -675,7 +417,7 @@ fn build_text_header(
         stream_id,
         timestamp: Utc::now().timestamp_millis(),
         topic: options.topic.clone(),
-        mime_type: TEXT_MIME_TYPE.to_owned(),
+        mime_type: constants::TEXT_MIME_TYPE.to_owned(),
         total_length,
         encryption_type: proto::encryption::Type::None.into(),
         attributes: options.attributes.clone(),
@@ -701,7 +443,7 @@ fn build_byte_header(
         stream_id,
         timestamp: Utc::now().timestamp_millis(),
         topic: options.topic.clone(),
-        mime_type: options.mime_type.clone().unwrap_or_else(|| BYTE_MIME_TYPE.to_owned()),
+        mime_type: options.mime_type.clone().unwrap_or_else(|| constants::BYTE_MIME_TYPE.to_owned()),
         total_length,
         encryption_type: proto::encryption::Type::None.into(),
         attributes: options.attributes.clone(),
@@ -713,19 +455,6 @@ fn build_byte_header(
     };
     (header, byte_header)
 }
-
-/// Max chunk content size AND the header-packet MTU budget. Kept below the ~16 KB
-/// data-channel MTU for protocol/E2EE framing headroom.
-const STREAM_CHUNK_SIZE_BYTES: usize = 15000;
-
-// Default MIME type to use for byte streams.
-static BYTE_MIME_TYPE: &str = "application/octet-stream";
-
-/// Default MIME type to use for text streams.
-static TEXT_MIME_TYPE: &str = "text/plain";
-
-/// Default name for `send_bytes` byte-stream headers.
-static BYTE_DEFAULT_NAME: &str = "unknown";
 
 #[cfg(test)]
 mod tests {
@@ -988,10 +717,10 @@ mod tests {
         assert!(h.inline_content.is_none());
         let chunks: Vec<_> = p[1..p.len() - 1].iter().map(chunk).collect();
         // Multi-packet, but fewer chunks than an uncompressed send would need (ceil(len/15000)).
-        let uncompressed_chunks = text.len().div_ceil(STREAM_CHUNK_SIZE_BYTES);
+        let uncompressed_chunks = text.len().div_ceil(constants::STREAM_CHUNK_SIZE_BYTES);
         assert!(chunks.len() >= 2);
         assert!(chunks.len() < uncompressed_chunks);
-        assert_eq!(chunks[0].content.len(), STREAM_CHUNK_SIZE_BYTES); // first chunk is full MTU
+        assert_eq!(chunks[0].content.len(), constants::STREAM_CHUNK_SIZE_BYTES); // first chunk is full MTU
         let total: usize = chunks.iter().map(|c| c.content.len()).sum();
         assert!(total < text.len()); // compressed
         assert_trailer(p.last().unwrap());
@@ -1308,7 +1037,7 @@ mod tests {
                 stream_id: "gc-test-stream".to_string(),
                 timestamp: 0,
                 topic: "gc-test-topic".to_string(),
-                mime_type: TEXT_MIME_TYPE.to_owned(),
+                mime_type: constants::TEXT_MIME_TYPE.to_owned(),
                 total_length: None,
                 encryption_type: proto::encryption::Type::None.into(),
                 attributes: HashMap::new(),
